backend/gateway/requirements.txt
fastapi
uvicorn
httpx
aioredis
setuptools
redis
PyJWT
